Added in saving and loading of Highscores
Highscores are now saved and loaded from disk, so they will persist even
after closing the game. Had to pull out ScoreProfile from HighScorePanel
to get this to work.
Also added in a restart level method which restarts the timer and resets
player position back to start.
Still need to test with a build made in unity.